Fix delelte

+++ src/application/use-case/language/command/delete-language/delete-language.command.ts
@@ -1,5 +1,5 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
-import { ForbiddenException, Inject } from "@nestjs/common";
+import { ForbiddenException, Inject, NotAcceptableException } from "@nestjs/common";
 import { Connection } from 'typeorm';
 import { PositionRepositoryOrm } from "@/infrastructures/repositories/position/position.repository";
 import { DeleteLanguageCommand } from "./delete-laguage.handler";
@@ -17,6 +17,8 @@ export class DeleteLanguageHandler implements ICommandHandler<DeleteLanguageComm
     async execute(command: DeleteLanguageCommand): Promise<void> {
         return await this.connection.transaction(async (manager) => {
         try{
+            const currentLanguage = await this.languageRepository.findById(command.id)
+            if(currentLanguage.languageProject.length > 0)throw new  NotAcceptableException({message:"Language is in project!"})
                 const language = await this.languageRepository.delete(command.id)
                 return language
             }catch(error){
