+++ src/application/use-case/position/command/delete-postion/delete-postion.handler.ts
@@ -1,5 +1,5 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
-import { ForbiddenException } from "@nestjs/common";
+import { ForbiddenException, NotAcceptableException } from "@nestjs/common";
 import { Connection } from 'typeorm';
 import { DeletePositionCommand } from "./delete-position.command";
 import { PositionRepositoryOrm } from "@/infrastructures/repositories/position/position.repository";
@@ -13,7 +13,9 @@ export class DeletePositionHandler implements ICommandHandler<DeletePositionComm
     }
     async execute(command: DeletePositionCommand): Promise<void> {
         return await this.connection.transaction(async (manager) => {
-        try{
+        try{    
+                const currrentPosition = await this.positionRepository.findById(command.id)
+                if(currrentPosition.positionMember.length >0) throw new NotAcceptableException({message: "Postion is in project!"})
                 const technical = await this.positionRepository.delete(command.id)
                 return technical
             }catch(error){
