+++ src/application/use-case/technical/command/delete-technical/delete-technical.handler.ts
@@ -1,6 +1,6 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
 // import { CreatePositionCommand } from "./create-position.command";
-import { ForbiddenException } from "@nestjs/common";
+import { ForbiddenException, NotAcceptableException } from "@nestjs/common";
 import { TechnicalM } from "@/domain/model/technical.model";
 import { TechnicalRepositoryOrm } from "@/infrastructures/repositories/technical/technical.repository";
 import { Connection } from 'typeorm';
@@ -16,6 +16,8 @@ export class DeleteTechnicalHandler implements ICommandHandler<DeleteTechnicalCo
     async execute(command: DeleteTechnicalCommand): Promise<void> {
         return await this.connection.transaction(async (manager) => {
         try{
+            const currentTechnical = await this.technicalRepository.findById(command.id)
+            if(currentTechnical.technicalProject.length > 0 ) throw new NotAcceptableException({message:'Technical is in project!'})
                 const technical = await this.technicalRepository.delete(command.id,manager)
                 return technical
             }catch(error){
